refactor and finish the remaining api endpoints
